Server function for module that makes a dygraph.

Used to define the server within the dygraph shiny module.

dygraph_server(input, output, session, data)

dygraph_sidebar_server(input, output, session, data)

Arguments

input

standard shiny input

output

standard shiny output

session

standard shiny session

data

data.frame or a reactive that returns a data.frame

Value

dygraph_server

a list containing:

rct_result a shiny::reactive, returning the resulting data.frame

rct_input_state a shiny::reactive, returning a list of logicals describing the state of the inputs

rct_status_content a shiny::reactive, returning a list with members class and message used to build the status output.

dygraph_sidebar_server

a reactive that returns a data.frame

See also

Other dygraph module functions: dygraph_sidebar_side, dygraph_ui_input, dygraph_ui_misc, dygraph_ui_output

Examples

library("shiny")
shinyServer(function(input, output, session) {

  list_rct <- callModule(
    module = dygraph_sidebar_server,
    id = "foo",
    data = shinypod::wx_ames
  )

})