Validates inputs for shiny modules, returns reactive

Using this function at the start of your Shiny module allows you to call your modules using arguments that may or may not be reactive.

reactive_validate(x, .f = identity, message, label = NULL, ...)

Arguments

x

expression, or shiny::reactive that returns an expression

.f

function that takes a single arg (object), returns TRUE if valid

message

character, passed to shiny::need

label

character, passed to shiny::need

...

other args passed along to .f

Value

shiny::reactive that returns x, or x(), if x is reactive

Details

This function does three things, but these things are associated as to form a pattern that merits its own function. For a given expression x:

if x is reactive, set y <- x(); if x is not reactive, set y <- x

validate y, using shiny::validate and shiny::need with .f, message, label, and ...

return shiny::reactive(y) to use later in your module

Examples

library("shiny")

# module function
summary_server <- function(input, output, session, df){

  rct_df <- reactive_validate(df, is.data.frame)

  rct_summary <- reactive(summary(rct_df()))

  rct_summary
}

# server function
shinyServer(function(input, output, session){

  # either of these constructions will work

  df <- mtcars
  rct_summary <- callModule(summary_server, "summary", df)

  rct_df <- reactive(mtcars)
  rct_summary <- callModule(summary_server, "summary", rct_df)

})