Constructs sp_srv object: template for server-return

This function can be useful for composing your own shinypods.

sp_srv(rct_result, rct_state, rct_notification)

Value

sp_srv object, S3 class

Details

The idea of the foundation layer is to define the inputs, outputs, and how they relate to each other. A sp_srv object has three members:

rct_result

A reactive that returns the result of the server function

rct_state

A reactive that returns a list of (presumably) logical values that can be used by the presentation server-function, for example, to show/hide various elements using shinyjs::toggle.

rct_notification

A reactive that contains a list of arguements that the presentation server-function can use to construct a notification using shiny::showNotification

Presumably, the presentation server-function itself return rct_result, act upon the information in rct_state to show or hide UI elements, and issue notifications based on rct_notification.

See also

Accessor functions: sp_rct_result.sp_srv, sp_rct_state.sp_srv, sp_rct_notification.sp_srv