Presentation UI-elements for upload-text module

These functions are used to create the UI elements for a sidebar presentation of the upload-text module.

upload_text_sb_side(id)

upload_text_sb_main(id)

Arguments

id

character, id for the module

Value

upload_text_sb_side

shiny::tagList of containing elements to put in a shiny::sidebarPanel

upload_text_sb_main

shiny::tagList of containing elements to put in a shiny::mainPanel

Details

upload_text_sb_side

The only element in the sidebar panel is a shiny::fileInput, used to upload a file from a local computer to the server.

upload_text_sb_main

The only element in the main panel is a shiny::htmlOutput, used to display the first few lines of the uploaded file.

See also

upload_text_ui, upload_text_sb_server

Examples

library("shiny")
library("shinyjs")
library("shinypod")

# also available by running sp_shiny_example("upload_text")

ui <- shinyUI(
  fluidPage(
    useShinyjs(),
    titlePanel("Upload Text"),
    sidebarLayout(
      sidebarPanel(
        upload_text_sb_side("upload")
      ),
      mainPanel(
        upload_text_sb_main("upload")
      )
    )
  )
)

server <- shinyServer(function(input, output, session) {

  # rct_text is a reactive that can be used elsewhere in a server function
  rct_text <- callModule(module = upload_text_sb_server, id = "upload")

})

## Not run: ------------------------------------
#   shinyApp(ui = ui, server = server)
## ---------------------------------------------