Create a shinypod from a template.

This function is used to write out a template function for a shinypod, both for the functional layer, in foo_pod.R, and a presentation layer, in foo_pod_sidebar.R

use_pod(name, description, is_blocking = TRUE, overwrite = FALSE)

Arguments

name

character, name to prepend to the filenames

description

character, completes the phrase "module that ..."

is_blocking

logical, indicates if this pod will have a blocking button

overwrite

logical, same usage as devtools; indicates if an existing file can be overwritten

Value

list of TRUE values

Details

This assumes that you are working in a directory in an R package; the files will be written to the R directory.